requirements.txt
flask==3.0.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
google-generativeai==0.3.2
beautifulsoup4==4.12.2
PyPDF2==3.0.1
matplotlib==3.8.2
Werkzeug==3.0.1
